requirements.txt
aiogram==2.22.2
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
soupsieve==2.5
urllib3==2.2.1
wikipedia==1.4.0
yarl==1.9.4
